backend/requirements.txt
Django==5.1
djangorestframework==3.15
django-cors-headers==4.6
Pillow==11.1
python-decouple==3.8
